feat: move ThreadSessionRestorer and ThreadRestorerDelegate to shared module (#4598)
